Add files via upload
Basic, slapped together script to fetch, configure, and run lbrycrdd. Tested on Ubuntu 18.04 and Windows 10 with python 3.7. Note on windows, the only reliable way I found to get it to work was to import the script as a module from a Python interrupter.  Will need further testing if the desired result is an end user simply clicking on the script to run it.